Move ElevationControl function to its own file
